ci: address review feedback on amber/amber-integration split
- labeler.yml: include amber/.scalafmt.conf and amber/.scalafix.conf
  under the engine label so a PR that only touches those configs still
  triggers the Scala stacks (instead of producing an empty stack
  union and skipping CI).
- labeler.yml: fix the amber-integration comment to describe the real
  layout (specs live under amber/src/test/integration/**, no /scala
  subdir) and the actual sbt wiring (Test/unmanagedSourceDirectories).
- IntegrationTest.java: rewrite the Javadoc to describe the real
  routing mechanism — ScalaTest tag filtering driven by
  AMBER_TEST_FILTER in amber/build.sbt — instead of the abandoned
  "dedicated IntegrationTest sbt configuration" wording.
- TestUtils: extract shouldReconfigure here, taking ActorSystem and
  WorkflowContext as params, and turn the per-spec defs in
  ReconfigurationSpec / ReconfigurationIntegrationSpec into thin
  wrappers, so future e2e specs reuse the helper without drift.
